# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateExperienciaRequest:
    properties:
      atividades:
        items:
          type: string
        type: array
      cargo:
        type: string
      empresa:
        type: string
      localizacao:
        type: string
      ordem:
        type: integer
      periodo:
        type: string
    type: object
...
  models.ExperienciaProfissional:
    properties:
      atividades:
        items:
          type: string
        type: array
      cargo:
        type: string
      empresa:
        type: string
      id:
        type: integer
      localizacao:
        type: string
      ordem:
        type: integer
      periodo:
        type: string
    type: object
...
  models.Projeto:
    properties:
      descricao:
        type: string
      tecnologias:
        items:
          type: string
        type: array
      titulo:
        type: string
    type: object
...
  models.UpdateExperienciaRequest:
    properties:
      atividades:
        items:
          type: string
        type: array
      cargo:
        type: string
      empresa:
        type: string
      localizacao:
        type: string
      ordem:
        type: integer
      periodo:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API REST para gerenciamento de portfólio pessoal
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Portfolio API
  version: "1.0"
paths:
  /api/admin/experiences:
    get:
      description: Retorna todas as experiências profissionais
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ExperienciaProfissional'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista experiências
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Cria uma nova experiência profissional
      parameters:
      - description: Dados da experiência
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateExperienciaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria experiência
      tags:
      - admin
  /api/admin/experiences/{id}:
    delete:
      description: Deleta uma experiência profissional
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deleta experiência
      tags:
      - admin
    get:
      description: Retorna uma experiência profissional por ID
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExperienciaProfissional'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtém experiência
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Atualiza uma experiência profissional existente
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: integer
      - description: Dados da experiência
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateExperienciaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza experiência
      tags:
      - admin
  /api/admin/experiences/{id}/translations:
    get:
      description: Retorna as traduções de uma experiência para os idiomas en e es
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/models.UpdateExperienciaRequest'
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtém traduções
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Salva as traduções de uma experiência para os 3 idiomas
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: integer
      - description: Traduções por idioma (pt-BR, en, es)
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            $ref: '#/definitions/models.UpdateExperienciaRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Salva traduções
      tags:
      - admin
  /api/admin/projects/{id}:
    put:
      consumes:
      - application/json
      description: Atualiza um projeto existente
      parameters:
      - description: ID do projeto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do projeto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Projeto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza projeto
      tags:
      - admin
  /api/admin/translations/{language}:
    put:
      consumes:
      - application/json
      description: Atualiza o valor de uma chave de tradução para um idioma específico
      parameters:
      - description: Código do idioma (pt-BR, en, es)
        in: path
        name: language
        required: true
        type: string
      - description: Chave e valor da tradução (key, value)
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza tradução
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"